+++ src/client/components/PlainMarkdownInput/plugins/slate-autocomplete-plugin/AutocompleteWidget.js
@@ -9,32 +9,25 @@ const maxHeight = 240;
 
 export default class AutocompleteWidget extends React.Component {
   static propTypes = {
-    isMouseIndexSelected: Types.bool,
-    onSelectedIndexChange: Types.func,
     items: Types.array,
+    loading: Types.bool,
     locale: Types.string,
-    onMouseDown: Types.func,
-    onMouseUp: Types.func,
-    onScroll: Types.func,
-    onSelectItem: Types.func,
-    selectedIndex: Types.number,
-    style: Types.object,
+    onChange: Types.func,
+    itemRenderer: Types.func,
     restrictorRef: Types.object,
-    renderItem: Types.func,
-    loading: Types.bool
+    selectedItem: Types.number,
+    style: Types.object
   }
 
   static defaultProps = {
-    isMouseIndexSelected: false,
-    onSelectedIndexChange: () => {},
     items: [],
+    loading: false,
     locale: 'en',
-    onMouseDown: () => {},
-    onMouseUp: () => {},
-    onScroll: () => {},
-    onSelectItem: () => {},
-    style: {},
-    restrictorRef: null
+    onChange: () => {},
+    itemRenderer: DefaultAutocompleteItem,
+    restrictorRef: null,
+    selectedItem: null,
+    style: {}
   };
 
   constructor(props) {
@@ -48,7 +41,6 @@ export default class AutocompleteWidget extends React.Component {
 
   componentDidMount = () => {
     this.adjustPosition();
-    this['items-ref'].addEventListener('scroll', this.props.onScroll, false);
   };
 
   componentWillReceiveProps = (nextProps) => {
@@ -57,16 +49,14 @@ export default class AutocompleteWidget extends React.Component {
   };
 
   componentWillUpdate = (nextProps) => {
-    let { isMouseIndexSelected } = this.props;
-    let itemsRef = this['items-ref'];
-    let itemRef = this[`item-ref-${nextProps.selectedIndex}`];
-
-    if (itemsRef && itemRef && !isMouseIndexSelected) { // calculating scrolling with keyboard up and down arrows
-      if ((this.props.selectedIndex < nextProps.selectedIndex) && (itemRef.offsetTop - itemsRef.scrollTop > 156)) {
-        itemsRef.scrollTop = (itemRef.offsetTop - 156);
+    let containerRef = this.containerRef;
+    let itemRef = this[`itemRef${nextProps.selectedItem}`];
+    if (containerRef && itemRef) { // calculating scrolling with keyboard up and down arrows
+      if ((this.props.selectedItem < nextProps.selectedItem) && (itemRef.offsetTop - containerRef.scrollTop > 156)) {
+        containerRef.scrollTop = (itemRef.offsetTop - 156);
       }
-      if ((this.props.selectedIndex > nextProps.selectedIndex) && (itemRef.offsetTop - itemsRef.scrollTop < 26)) {
-        itemsRef.scrollTop = (itemRef.offsetTop - 26);
+      if ((this.props.selectedItem > nextProps.selectedItem) && (itemRef.offsetTop - containerRef.scrollTop < 26)) {
+        containerRef.scrollTop = (itemRef.offsetTop - 26);
       }
     }
   };
@@ -81,15 +71,11 @@ export default class AutocompleteWidget extends React.Component {
     }
   };
 
-  handleSelectItem = index => {
-    this.props.onSelectItem(index);
-  };
-
-  setPosition = (selection) => {
+  setPosition = (selectedItem) => {
     let editorWidth = this.props.restrictorRef.offsetWidth;
-    let autocompleteWidth = this['items-ref'].offsetWidth;
-    let autocompleteHeight = this['items-ref'].offsetHeight;
-    let selectionRect = selection.getRangeAt(0).getBoundingClientRect(); // element with cursor
+    let autocompleteWidth = this['containerRef'].offsetWidth;
+    let autocompleteHeight = this['containerRef'].offsetHeight;
+    let selectionRect = selectedItem.getRangeAt(0).getBoundingClientRect(); // element with cursor
     let restrictorRect = this.props.restrictorRef.getBoundingClientRect();
     let lineHeight = selectionRect.bottom - selectionRect.top;
 
@@ -98,8 +84,8 @@ export default class AutocompleteWidget extends React.Component {
     left = left < 0 ? 0 : left;
 
     let top = selectionRect.top - restrictorRect.top + lineHeight + 4;
-    let offsetTop = selection.anchorNode.parentNode.offsetTop;
-    let slateEditor = getSlateEditor(selection);
+    let offsetTop = selectedItem.anchorNode.parentNode.offsetTop;
+    let slateEditor = getSlateEditor(selectedItem);
 
     slateEditor.style.overflow = 'hidden';
 
@@ -123,29 +109,15 @@ export default class AutocompleteWidget extends React.Component {
   };
 
   adjustPosition = () => {
-    let selection = window.getSelection();
+    let selectedItem = window.getSelection();
 
-    if (selection.anchorNode) {
-      this.setPosition(selection);
+    if (selectedItem.anchorNode) {
+      this.setPosition(selectedItem);
     }
   };
 
-  saveRef = ref => (this['items-ref'] = ref);
-
-  handleMouseDown = e => {
-    e.preventDefault(); // XXX Not work in IE11
-    e.stopPropagation(); // Isolate event target
-    this.props.onMouseDown('widget');
-  }
-
-  handleMouseUp = e => {
-    e.stopPropagation(); // Isolate event target
-    this.props.onMouseUp();
-  }
-
-  handleItemMouseDown = e => {
-    e.stopPropagation(); // Isolate event target
-    this.props.onMouseDown('item');
+  handleItemMouseDown = (index) => {
+    this.props.onChange(index);
   }
 
   render() {
@@ -153,65 +125,67 @@ export default class AutocompleteWidget extends React.Component {
     const {
       items,
       locale,
-      selectedIndex,
-      onSelectedIndexChange,
-      onSelectItem,
-      renderItem: CustomItemComponent,
+      selectedItem,
+      itemRenderer,
       loading
     } = this.props;
 
-    const ItemComponent = CustomItemComponent ? CustomItemComponent : DefaultAutocompleteItem;
-
-    const commonProps = {
-      className: "react-markdown--autocomplete-widget",
-      ref: this.saveRef,
-      style: {
-        left,
-        top,
-        transform,
-        maxHeight: `${maxHeight}px`,
-        ...this.props.style
-      }
-    }
-
     if (loading) {
       return (
-        <div {...commonProps}>
+        <div
+          ref={ref => (this.containerRef = ref)}
+          style={{
+            left,
+            top,
+            transform,
+            maxHeight: `${maxHeight}px`,
+            ...this.props.style
+          }}
+          className="react-markdown--autocomplete-widget"
+        >
           <div className="react-markdown--autocomplete-widget__item">
             <span>{getMessage(locale, 'searching')}</span>
             <i className="fa fa-spinner fa-spin pull-right" style={{ marginTop: '3px' }}></i>
           </div>
         </div>
-      )
+      );
     }
+
+    let notFoundElement = items.length === 0 && !loading ? (
+      <div className="react-markdown--autocomplete-widget__item">{getMessage(locale, 'noMatchesFound')}</div>
+    ) : null;
+
+    let itemsElement = items.map((item, index) => {
+      const isSelected = selectedItem === index;
+      return (
+        <div
+          key={index}
+          ref={ref => (this[`itemRef${index}`] = ref)}
+          onMouseDown={_ => this.handleItemMouseDown(index)}
+        >
+          <input type="radio" style={{ display: 'none' }} selected={isSelected} />
+          {itemRenderer({ item, isSelected })}
+        </div>
+      );
+    });
 
     if (items) {
       return (
         <div
-          {...commonProps}
-          onMouseDown={this.handleMouseDown}
-          onMouseUp={this.handleMouseUp}
+          ref={ref => (this.containerRef = ref)}
+          style={{
+            left,
+            top,
+            transform,
+            maxHeight: `${maxHeight}px`,
+            ...this.props.style
+          }}
+          className="react-markdown--autocomplete-widget"
         >
-          {items.map((item, index) => {
-            return (
-              <div
-                key={index}
-                ref={ref => (this[`item-ref-${index}`] = ref)}
-                onClick={_ => onSelectItem(index)}
-                onMouseMove={_ => onSelectedIndexChange(index)}
-                onMouseDown={this.handleItemMouseDown}
-              >
-                <ItemComponent
-                  item={item}
-                  isSelected={selectedIndex === index}
-                />
-              </div>
-            );
-          })}
-
-          {items.length === 0 && !loading && (
-            <div className="react-markdown--autocomplete-widget__item">{getMessage(locale, 'noMatchesFound')}</div>
-          )}
+          <fieldset>
+            {itemsElement}
+          </fieldset>
+          {notFoundElement}
         </div>
       );
     }
@@ -219,4 +193,3 @@ export default class AutocompleteWidget extends React.Component {
     return null;
   }
 }
-